requirements.txt
# File: requirements.txt
flask
yt-dlp
pillow